Update post api

+++ index.js
@@ -24,7 +24,7 @@ var wallet;
 
 function payToAddress(args, callback) {
 	var address = args.address
-	var amount = parseFloat(args.amount)
+	var amount = parseInt(args.amount)
 	console.log('===payTo: ' + address + " amount: " + amount + '===')
 	if(!ValidationUtils.isValidAddress(address)){
 		return callback('Uncorrect Address')
